Refactor blog data management: migrated blog posts and categories to a centralized data file for improved organization and maintainability, enhancing the structure of the blog component.
